requirements.txt
insights-core>=3.0.24
